Split test pkg into unit and integration test methods
gps is allowed to reference internal/test, but only dep should reference
internal/test/integration.
